usr: 64 bit dev_t

+++ usr.sbin/lpr/lpr/lpr.c
@@ -386,8 +386,8 @@ main(int argc, char *argv[])
 			continue;	/* file unreasonable */
 
 		if (sflag && (cp = linked(arg)) != NULL) {
-			(void) snprintf(buf, sizeof(buf), "%u %u", statb.st_dev,
-				statb.st_ino);
+			(void) snprintf(buf, sizeof(buf), "%ju %ju",
+			    (uintmax_t)statb.st_dev, (uintmax_t)statb.st_ino);
 			card('S', buf);
 			if (format == 'p')
 				card('T', title ? title : arg);
